% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agronomy.R
\name{fertilizer.dose}
\alias{fertilizer.dose}
\title{Calculate fertilizer dose for a given crop}
\usage{
fertilizer.dose(
  nitrogen,
  phosphorus,
  potassium,
  area = 1,
  DAP = FALSE,
  TSP = FALSE,
  complete = FALSE
)
}
\arguments{
\item{nitrogen}{Numeric, the recommended rate of Nitrogen in Kg/Ha(Kilogram per hectare) required for a crop.}

\item{phosphorus}{Numeric, the recommended rate of Phosphorous in Kg/Ha(Kilogram per hectare) required for a crop.}

\item{potassium}{Numeric, the recommended rate of Potassium in Kg/Ha(Kilogram per hectare) required for a crop.}

\item{area}{Numeric, the area of the land (in hectares) where the crop will be grown. Area is assumed to be 1 hectare by default}

\item{DAP}{Logical, whether DAP is to be used as a source of phosphate and nitrogen or not. DAP is assumed to be FALSE by default}

\item{TSP}{Logical, whether TSP is to be used as a source of phosphate or not. TSP is assumed to be FALSE by default}

\item{complete}{Logical, whether complete(NPK mix) is to be used as fertilizers (NPK) or not. Complete is assumed to be FALSE by default}
}
\value{
A numeric vector with the amount of urea, DAP, TSP, SSP, and muriate of potash (in kg) required for the crop.
}
\description{
This function calculates the amount of nitrogen, phosphate, and potash fertilizers required for a given crop based on the recommended rate of fertilizer usage. It supports different fertilizer sources including complete,muriate of potash, Diammonium Phosphate (DAP), Triple Super Phosphate (TSP), Single Super Phosphate (SSP),and Urea.
}
\examples{
# Calculate fertilizer dose using Urea, SSP and Muriate of potash for 1 hectare of cultivation area.
fertilizer.dose(120,60,60)

# Calculate fertilizer dose using Urea, SSP and Muriate of potash for 2 hectare of cultivation area.
fertilizer.dose(120,60,60, 2)

# Calculate fertilizer dose using DAP as a source of phosphorus and nitrogen.
fertilizer.dose(nitrogen = 100, phosphorus = 50, potassium = 75, area = 2, DAP = TRUE)

# Calculate fertilizer dose using TSP as a source of phosphorus.
fertilizer.dose(nitrogen = 100, phosphorus = 50, potassium = 75, area = 2, TSP = TRUE)

# Calculate fertilizer dose using a complete fertilizer.
fertilizer.dose(nitrogen = 100, phosphorus = 50, potassium = 75, area = 2, complete = TRUE)

}
