% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agronomy.R
\name{herbicide.rate}
\alias{herbicide.rate}
\title{Function to calculate the herbicide rate to be applied for a given crop}
\usage{
herbicide.rate(recommended_rate, area = 1, active_ingredients)
}
\arguments{
\item{recommended_rate}{Numeric, the recommended rate of herbicide application in Kg/Liter a.i. per ha (Kilogram active ingredients per hectare).}

\item{area}{Numeric, the area of the ground to be sprayed by the herbicide in hectares(ha).}

\item{active_ingredients}{Numeric, the percentage of active ingredients in percentage.}
}
\value{
The required amount of herbicide in liters for the given area.
}
\description{
Calculates the required amount of herbicide for a given area, based on the recommended rate and active ingredients.
}
\examples{
#calculate the herbicide when 2 kilogram active ingredients per hectare is recommended for one hectare of land with 35 \% active ingredients)
herbicide.rate(2, 1, 35)
}
