% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agronomy.R
\name{seed.rate}
\alias{seed.rate}
\title{Function to calculate the seed rate and plant population for a given crop}
\usage{
seed.rate(
  area = 1,
  TSW,
  row_spacing,
  plant_spacing,
  germination = 100,
  purity = 100,
  plant_per_hill = 1,
  gap_filling = 0
)
}
\arguments{
\item{area}{Numeric, the area to be sown in hectares(ha).}

\item{TSW}{Numeric, the thousand seed weight in gram(gm) of a given plant.}

\item{row_spacing}{Numeric, the spacing between rows in centimeter(cm).}

\item{plant_spacing}{Numeric, the spacing between Plants in centimeter(cm).}

\item{germination}{Numeric, the germination rate of the seed in percentage, 100 percent germination is considered by default.}

\item{purity}{Numeric, the purity rate of the seed in percentage, 100 percent purity is considered by default.}

\item{plant_per_hill}{Numeric, the desired number of plants per hill, one plant per hill is considered by default.}

\item{gap_filling}{Numeric, the percentage of seed required for gap filling.}
}
\value{
A string with the calculated seed rate and plant population.
}
\description{
this function calculates the seed rate and plant population required for a given field area, based on the total seed weight (TSW), row spacing, plant spacing, germination rate, purity rate, and desired number of plants per hill.
}
\examples{
# Calculation of seed rate when area, TSW, row_spacing, plant_spacing, germination, and purity is given.
seed.rate(1, 75, 60,30, 90, 85)

# Calculation of seed rate when two seeds are sown per hills.
seed.rate(area = 2, TSW = 76, row_spacing = 60, plant_spacing = 30, germination = 96, purity = 95, plant_per_hill = 2)

# Calculation of seed rate when shelling is 80\% and gap filling is 10\%.
seed.rate(area = 1, TSW = 90, row_spacing = 75, plant_spacing = 25, germination = 95, purity = 95, plant_per_hill = 1, gap_filling=10)

}
